Move rom to mbc module
