+++ crates/memory/src/mbc/mbc5.rs
@@ -1,5 +1,5 @@
-use super::consts;
 use super::bus::MbcBus;
+use super::consts;
 use crate::MemoryBus;
 use shared::Error;
